After an event with Nancy and clinical, I made the navigation work VERBATIM as the Flash engine.
The navigation can be closed and opened by clicking on the sub navigation landing page, HOWEVER if you navigate to a page INSIDE the sub navigation section, the sub navigation will be opened. The previous iteration meant the pages could not be collapsible (good thing I knew to comment out the code instead of deleting it as per requested...).

There were some other minor alterations like decreasing the size of the font, making the bottom border extend all the way to the left, and as a bonus, I got to use the illusive adjacent sibling operator, +. 

Also made the border color between navigation items and subnav arrow use the productPreferences colors.