These changes all stem from VUE.js being used to set DepartmentID.
Because field does not exist on page load, I need to post to the page with the variables. Problem was I needed a cookie in order to have session information. $this has a response with the cookie. We now use that cookie to post to page and everything works. Removed logoff because it clears the cookie.

https://s-media-cache-ak0.pinimg.com/originals/59/72/d6/5972d6bcc10bca103fffaac1184c3431.jpg