Beta users should see all products regardless of their active/beta status.
If a product is not active but users have licenses, they can still see product. All of our employee account are set to Beta so we can see edit info in courses. If the product was turned off and not in beta, then we could not see them. So, remove section which blocked the products.